Copy stuff from Gulpfile.js into submodules
